ports usually come as ints

+++ multi-file-server.js
@@ -6,7 +6,7 @@ var sys = require("sys"),
     multi = require("multi-node");
 
 var dir = process.argv[2] || './public';
-var port = parseFloat(process.argv[3]) || 8080;
+var port = parseInt(process.argv[3]) || 8080;
 sys.log('Serving files from ' + dir + ', port is ' + port);
 
 var server = http.createServer(function(request, response) {
